A simple Shallow Neural Net made with Keras for MNIST
It achieves 98.4% accuracy and converges pretty fast as "ADAM" optimizer is used for learning and "relu" made it significantly faster than "sigmoid". However if you want to test with "sigmoid" or "tanh" (does not matter as the next layer is softmax, so only biases will change mostly) you can just change "relu" in the code to "sigmoid" or "tanh"or any other function from the pre-provided list in keras documentation.