Shallow CNN on MNIST, ~99% validation accuracy
We have used two Conv layers and 1 maxpooling layer, in the fully connected layer we used a shallow NN with 128 hidden nodes. We used "relu" throughout as the activation function for faster convergence and "adadelta" optimizer for the same reason. After Maxpooling 50% dropout is used and in the final hidden layer 30%. The last maxpooling layer is a stack of most important indicators for the image. By introducing high dropout, we ensure that any one indicator does not become the determiner in any case.